SpellsLvl01: Use size_t

+++ src/game/magic/spells/SpellsLvl01.h
@@ -72,7 +72,7 @@ private:
 	
 	struct T_LINKLIGHTTOFX {
 		LightHandle m_effectLight;
-		int m_targetLight;
+		size_t m_targetLight;
 	};
 	
 	std::vector<T_LINKLIGHTTOFX> m_lights;
